+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+# 2023-02-17
+
+* Add support for turning off the heater entirely. The last preferred target temperature is persisted to local device storage.
+  Fixes issue #1.
+
 # 2023-02-16
 
 * Add support for temperature unit. User can switch between Celsius and Fahrenheit. The setting is persisted in the mug.
-  ffcea29083e86d06ba4a3ae99ec5a4a1bb021575. Fixes #2.
+  Fixes issue #2.